+++ uefi/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Added
 - Added `proto::console::text::InputEx`.
+- Added `proto::pci::PciRootBridgeIo::{supported_attributes(), attributes(),
+  set_attributes(), set_attributes_with_range()}`
 
 ## Changed
 - MSRV increased from 1.88 to 1.91.